Convert `PointerTensor` into a generic `Pointer` class
The fields on the pointer types are all subsets of the same set of
fields. Capturing them as separate schemas might offer some additional
type safety, but feels unnecessarily repetitive and prone to maintenance
issues. They can be split later if there's a concrete reason to do so,
but seems like overkill for now. (Adding a field to capture the object
type of the pointer as a substitute.)